tests/wpt/tests/fetch/metadata/generated/worker-dedicated-constructor.sub.html

<!DOCTYPE html>
<!--
This test was procedurally generated. Please do not modify it directly.
Sources:
- fetch/metadata/tools/fetch-metadata.conf.yml
- fetch/metadata/tools/templates/worker-dedicated-constructor.sub.html
-->
<html lang="en">
  <meta charset="utf-8">
  <title>HTTP headers on request for dedicated worker via the "Worker" constructor</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/fetch/metadata/resources/helper.sub.js"></script>
  <script type="module">
  'use strict';
  function induceRequest(url, options) {
    return new Promise((resolve, reject) => {
      const worker = new Worker(url, options);
      worker.onmessage = resolve;
      worker.onerror = reject;
    });
  }

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key,
      ['httpOrigin'],
      { mime: 'application/javascript', body: 'postMessage("")' }
    );

    return induceRequest(url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-site');
        });
  }, 'sec-fetch-site - Not sent to non-trustworthy same-origin destination, no options');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key,
      ['httpOrigin'],
      { mime: 'application/javascript', body: 'postMessage("")' }
    );

    return induceRequest(url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-mode');
        });
  }, 'sec-fetch-mode - Not sent to non-trustworthy same-origin destination, no options');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key,
      ['httpOrigin'],
      { mime: 'application/javascript', body: 'postMessage("")' }
    );

    return induceRequest(url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-dest');
        });
  }, 'sec-fetch-dest - Not sent to non-trustworthy same-origin destination, no options');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key,
      ['httpOrigin'],
      { mime: 'application/javascript', body: 'postMessage("")' }
    );

    return induceRequest(url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-user');
        });
  }, 'sec-fetch-user - Not sent to non-trustworthy same-origin destination, no options');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key,
      ['httpOrigin'],
      { mime: 'application/javascript', body: 'postMessage("")' }
    );

    return induceRequest(url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-storage-access');
        });
  }, 'sec-fetch-storage-access - Not sent to non-trustworthy same-origin destination, no options');
  </script>
</html>
